updata committer adivce

Created-by: liumingxuan9
Commit-by: @liu_mingxuan
Merged-by: liumingxuan9
Description: ## 描述
<!--在这里详细描述你的改动，包括改动的原因和所采取的方法。-->

## 关联的Issue
<!-- 如果这个PR是为了解决特定的Issue，请在这里提供Issue链接。例如：关联Issue #000-->
<!-- 如果这个PR是为了解决特定的问题单，请在这里描述问题单单号。-->

## 测试
<!--描述进行了哪些测试来验证你的改动。包括但不限于二级冒烟、算子泛化等。-->

## 文档更新
<!--如果这个PR包含文档的更新，请在这里指出。例如：更新了README.md文件。-->

## 类型标签
<!--  [x] 表示选中 -->
- [ ] 🐛 Bug 修复
- [ ] ✨ 新特性
- [ ] ⚡ 性能优化
- [ ] ♻️ 重构
- [ ] 🧪 测试
- [ ] 📦 构建/CI
- [ ] 🔧 配置变更
- [ ] 📝 文档更新
- [ ] ⬆️ 依赖升级
- [ ] 🔒 安全修复
- [ ] 🧹 代码清理
- [ ] ❓ 其他，请描述：



See merge request: l00858142/ops-transformer!8

+++ mc2/moe_distribute_dispatch_v2/op_kernel/moe_distribute_v2_constant.h
@@ -75,9 +75,9 @@ constexpr uint32_t PERGROUP_DYNAMIC_QUANT = 3;
 constexpr uint32_t MX_QUANT = 4;
 constexpr uint32_t CACHEWRITESIZE = 8U;
 
-constexpr uint32_t INT8_COMM_QUANT = 2;
-constexpr uint32_t MXFP8_E5M2_COMM_QUANT = 3;
-constexpr uint32_t MXFP8_E4M3_COMM_QUANT = 4;
+constexpr uint8_t INT8_COMM_QUANT = 2;
+constexpr uint8_t MXFP8_E5M2_COMM_QUANT = 3;
+constexpr uint8_t MXFP8_E4M3_COMM_QUANT = 4;
 
 // combinev2额外所需常量段
 constexpr uint64_t COMBINE_STATE_WIN_OFFSET = 818UL * 1024UL;   // combine 0/1状态区偏移，为dispatch预留50k大小
